Build a scalable layered test harness
