+++ package.json
@@ -22,15 +22,15 @@
     "dev:environment:logs": "node scripts/dev-environment.mjs logs",
     "dev:desktop:isolated": "node scripts/isolated-desktop.mjs",
     "dev:desktop:fresh": "node scripts/isolated-desktop.mjs --fresh",
-    "e2e": "pnpm build && cargo build --workspace && node scripts/e2e.mjs",
-    "e2e:container": "node scripts/container-e2e.mjs",
-    "e2e:desktop:container": "pnpm --filter @mdbase/connect-desktop build:standalone && cargo build -p mdbase-cli && node apps/desktop/scripts/docker-server-e2e.mjs",
-    "e2e:ecosystem": "pnpm e2e:container && pnpm e2e:desktop:container",
-    "e2e:relay": "pnpm build && node scripts/relay-e2e.mjs",
-    "e2e:sync": "pnpm build && node scripts/sync-e2e.mjs",
-    "e2e:provider": "pnpm build && cargo build -p mdbase-connect-hosted-provider -p mdbase-cli && node scripts/hosted-provider-e2e.mjs",
-    "e2e:files": "pnpm build && cargo build -p mdbase-connect-hosted-provider && node scripts/hosted-files-e2e.mjs",
-    "e2e:files:adversarial": "node scripts/hosted-file-adversarial-e2e.mjs",
+    "e2e": "node test/system/run.mjs --suite local",
+    "e2e:container": "node test/system/run.mjs --suite container",
+    "e2e:desktop:container": "node test/system/run.mjs --suite desktop",
+    "e2e:ecosystem": "node test/system/run.mjs --suite container,desktop",
+    "e2e:relay": "node test/system/run.mjs --suite relay",
+    "e2e:sync": "node test/system/run.mjs --suite sync",
+    "e2e:provider": "node test/system/run.mjs --suite provider",
+    "e2e:files": "node test/system/run.mjs --suite files",
+    "e2e:files:adversarial": "node test/system/run.mjs --suite files-adversarial",
     "e2e:authority:stress": "MDBASE_CONNECT_PROVIDER_E2E_PROMOTION_COUNT=2500 pnpm e2e:provider",
     "e2e:provider:stress": "MDBASE_CONNECT_PROVIDER_E2E_BULK_COUNT=10000 pnpm e2e:provider",
     "profile:mirror": "node --expose-gc scripts/mirror-profile.mjs",
@@ -53,6 +53,10 @@
     "test:stress-regressions": "pnpm stress:prepare && node scripts/stress/replay-regressions.mjs",
     "pretest": "pnpm build:packages",
     "test": "node --test scripts/lib/*.test.mjs && pnpm --filter @mdbase/connect-desktop test && pnpm --filter '!@mdbase/connect-desktop' -r test",
+    "test:fast": "pnpm test && node test/rust/run.mjs",
+    "test:integration": "pnpm test:browser-storage && pnpm test:accessibility",
+    "test:system": "node test/system/run.mjs",
+    "test:all": "pnpm test:fast && pnpm test:integration && pnpm test:system",
     "test:accessibility": "pnpm build:packages && pnpm --filter @mdbase/connect-portal build && pnpm --filter @mdbase/connect-desktop build:renderer && pnpm --filter mdbase-editor build && node scripts/browser-accessibility.test.mjs",
     "test:browser-storage": "pnpm --filter @mdbase-dev/connect test:browser-storage",
     "pretypecheck": "pnpm build:packages",
