+++ scripts/lib/connect-test-environment.test.mjs
@@ -1,6 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
+import { createConnectEnvironment } from "./connect-environment.mjs";
 import { sanitizeProjectName } from "./connect-test-environment.mjs";
 
 test("keeps already-valid Compose project names stable", () => {
@@ -26,5 +27,33 @@ test("preserves uniqueness when normalization changes project names", () => {
   assert.notEqual(
     sanitizeProjectName("desktop suite"),
     sanitizeProjectName("desktop-suite")
+  );
+});
+
+test("configures persistent development environments without ephemeral secrets", async () => {
+  const environment = await createConnectEnvironment({
+    projectName: "mdbase-connect-dev",
+    connectPort: 18787,
+    natsPort: 14222,
+    build: true,
+    disposable: false,
+    randomizeCredentials: false,
+    environment: {
+      PUBLIC_URL: "http://127.0.0.1:18787",
+      MDBASE_EDITOR_ORIGIN: "http://127.0.0.1:5173"
+    }
+  });
+
+  assert.equal(environment.projectName, "mdbase-connect-dev");
+  assert.equal(environment.serverUrl, "http://127.0.0.1:18787");
+  assert.equal(environment.environment.MDBASE_CONNECT_BIND_PORT, "18787");
+  assert.equal(environment.environment.MDBASE_CONNECT_NATS_BIND_PORT, "14222");
+  assert.equal(
+    environment.environment.MDBASE_EDITOR_ORIGIN,
+    "http://127.0.0.1:5173"
+  );
+  assert.equal(
+    Object.hasOwn(environment.environment, "POSTGRES_PASSWORD"),
+    Object.hasOwn(process.env, "POSTGRES_PASSWORD")
   );
 });
