+++ scripts/lib/upgrade-test-contract.test.mjs
@@ -1,0 +1,64 @@
+import assert from "node:assert/strict";
+import { execFile } from "node:child_process";
+import { readFile } from "node:fs/promises";
+import { resolve } from "node:path";
+import test from "node:test";
+import { promisify } from "node:util";
+import { availableTcpPort, poll } from "./test-runtime.mjs";
+
+const execute = promisify(execFile);
+const repoRoot = resolve(import.meta.dirname, "../..");
+
+test("upgrade workflows delegate scenario behavior to versioned test programs", async () => {
+  const workflow = await readFile(
+    resolve(repoRoot, ".github/workflows/server-ci.yml"),
+    "utf8"
+  );
+  assert.match(workflow, /run: test\/upgrade\/server-from-previous/);
+  assert.match(workflow, /run: test\/upgrade\/provider-from-previous/);
+  assert.doesNotMatch(workflow, /node --input-type=module --eval/);
+  assert.doesNotMatch(workflow, /INSERT INTO hosted_provider_/);
+});
+
+test("upgrade shell programs are syntactically valid", async () => {
+  for (const script of [
+    "test/upgrade/lib.sh",
+    "test/upgrade/server-from-previous",
+    "test/upgrade/provider-from-previous"
+  ]) {
+    await execute("bash", ["-n", resolve(repoRoot, script)]);
+  }
+});
+
+test("previous-provider fixture preserves the notification contract", async () => {
+  const fixture = await readFile(
+    resolve(repoRoot, "test/upgrade/provider-notification.sql"),
+    "utf8"
+  );
+  assert.match(fixture, /INSERT INTO hosted_provider_collections/);
+  assert.match(fixture, /INSERT INTO hosted_provider_notification_grants/);
+  assert.match(fixture, /mdbase\.runtime\.timer\.fired/);
+  assert.match(fixture, /"version":"1\.0\.0"/);
+});
+
+test("S3 readiness fixture serves a scoped empty bucket listing", async (context) => {
+  const port = await availableTcpPort();
+  const child = execFile(
+    process.execPath,
+    [resolve(repoRoot, "test/upgrade/r2-readiness-stub.mjs")],
+    { env: { ...process.env, R2_STUB_PORT: String(port) } }
+  );
+  context.after(() => child.kill("SIGTERM"));
+
+  const response = await poll(
+    () => fetch(`http://127.0.0.1:${port}`).catch(() => undefined),
+    "S3 readiness fixture did not start",
+    40,
+    25
+  );
+  assert.equal(response.status, 200);
+  const body = await response.text();
+  assert.match(body, /<Name>upgrade-canary<\/Name>/);
+  assert.match(body, /<Prefix>v1\/<\/Prefix>/);
+  assert.match(body, /<KeyCount>0<\/KeyCount>/);
+});
